Remove deprecated aliases and cleanup artifacts from `:memory` consolidation
PiperOrigin-RevId: 530640656
