+++ base/internal/function_adapter_test.cc
@@ -22,7 +22,7 @@
 #include "absl/time/time.h"
 #include "base/handle.h"
 #include "base/kind.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "base/type_factory.h"
 #include "base/type_manager.h"
 #include "base/type_provider.h"
