--- base/memory_manager.h
@@ -1,22 +1,0 @@
-// Copyright 2022 Google LLC
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     https://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-#ifndef THIRD_PARTY_CEL_CPP_BASE_MEMORY_MANAGER_H_
-#define THIRD_PARTY_CEL_CPP_BASE_MEMORY_MANAGER_H_
-
-// TODO(issues/5): delete
-
-#include "base/memory.h"
-
-#endif  // THIRD_PARTY_CEL_CPP_BASE_MEMORY_MANAGER_H_
