+++ base/type_factory.h
@@ -24,7 +24,7 @@
 #include "absl/status/statusor.h"
 #include "absl/synchronization/mutex.h"
 #include "base/handle.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "base/types/any_type.h"
 #include "base/types/bool_type.h"
 #include "base/types/bytes_type.h"
