+++ base/type_test.cc
@@ -22,7 +22,7 @@
 #include "absl/status/status.h"
 #include "base/handle.h"
 #include "base/internal/memory_manager_testing.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "base/type_factory.h"
 #include "base/type_manager.h"
 #include "base/value.h"
