+++ base/value_factory.h
@@ -31,7 +31,7 @@
 #include "base/attribute_set.h"
 #include "base/function_result_set.h"
 #include "base/handle.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "base/owner.h"
 #include "base/type_manager.h"
 #include "base/value.h"
