+++ base/value_factory_test.cc
@@ -15,7 +15,7 @@
 #include "base/value_factory.h"
 
 #include "absl/status/status.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "internal/testing.h"
 
 namespace cel {
