+++ base/values/map_value_builder.h
@@ -25,7 +25,7 @@
 #include "absl/base/attributes.h"
 #include "absl/base/macros.h"
 #include "absl/status/statusor.h"
-#include "base/allocator.h"
+#include "base/memory.h"
 #include "base/value_factory.h"
 #include "base/values/list_value_builder.h"
 #include "base/values/map_value.h"
