+++ eval/eval/BUILD
@@ -20,7 +20,7 @@ cc_library(
         ":evaluator_stack",
         "//base:ast_internal",
         "//base:handle",
-        "//base:memory_manager",
+        "//base:memory",
         "//base:type",
         "//base:value",
         "//eval/internal:adapter_activation_impl",
@@ -71,11 +71,8 @@ cc_test(
     ],
     deps = [
         ":evaluator_stack",
-        "//base:type_factory",
-        "//base:type_manager",
-        "//base:type_provider",
+        "//base:type",
         "//base:value",
-        "//base:value_factory",
         "//extensions/protobuf:memory_manager",
         "//internal:testing",
     ],
@@ -123,7 +120,7 @@ cc_library(
         ":expression_step_base",
         "//base:attributes",
         "//base:kind",
-        "//base:memory_manager",
+        "//base:memory",
         "//base:value",
         "//eval/internal:errors",
         "//eval/internal:interop",
@@ -230,7 +227,7 @@ cc_library(
         ":expression_step_base",
         "//base:ast_internal",
         "//base:handle",
-        "//base:memory_manager",
+        "//base:memory",
         "//base:type",
         "//base:value",
         "//eval/internal:errors",
@@ -668,7 +665,7 @@ cc_library(
     srcs = ["attribute_trail.cc"],
     hdrs = ["attribute_trail.h"],
     deps = [
-        "//base:memory_manager",
+        "//base:memory",
         "//eval/public:cel_attribute",
         "//eval/public:cel_expression",
         "//eval/public:cel_value",
@@ -709,7 +706,7 @@ cc_library(
         "//base:function_result",
         "//base:function_result_set",
         "//base:handle",
-        "//base:memory_manager",
+        "//base:memory",
         "//base:value",
         "//eval/public:unknown_set",
         "//extensions/protobuf:memory_manager",
