+++ eval/internal/adapter_activation_impl.cc
@@ -18,7 +18,7 @@
 
 #include "absl/strings/string_view.h"
 #include "absl/types/optional.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "eval/internal/interop.h"
 #include "eval/public/cel_value.h"
 #include "extensions/protobuf/memory_manager.h"
