+++ eval/public/BUILD
@@ -74,7 +74,7 @@ cc_library(
         ":message_wrapper",
         ":unknown_set",
         "//base:kind",
-        "//base:memory_manager",
+        "//base:memory",
         "//eval/internal:errors",
         "//eval/public/structs:legacy_type_info_apis",
         "//extensions/protobuf:memory_manager",
@@ -701,7 +701,7 @@ cc_test(
         ":cel_value_internal",
         ":unknown_attribute_set",
         ":unknown_set",
-        "//base:memory_manager",
+        "//base:memory",
         "//eval/internal:errors",
         "//eval/public/structs:legacy_type_info_apis",
         "//eval/public/structs:trivial_legacy_type_info",
