+++ extensions/protobuf/struct_value.cc
@@ -36,8 +36,8 @@
 #include "absl/time/time.h"
 #include "absl/types/optional.h"
 #include "absl/types/variant.h"
-#include "base/allocator.h"
 #include "base/handle.h"
+#include "base/memory.h"
 #include "base/types/struct_type.h"
 #include "base/value.h"
 #include "base/value_factory.h"
