+++ extensions/protobuf/type_provider_test.cc
@@ -16,7 +16,7 @@
 
 #include "google/protobuf/type.pb.h"
 #include "base/kind.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "base/type_factory.h"
 #include "base/type_manager.h"
 #include "extensions/protobuf/enum_type.h"
