+++ runtime/activation_test.cc
@@ -23,7 +23,7 @@
 #include "base/function.h"
 #include "base/function_descriptor.h"
 #include "base/handle.h"
-#include "base/memory_manager.h"
+#include "base/memory.h"
 #include "base/type_factory.h"
 #include "base/type_manager.h"
 #include "base/type_provider.h"
